1.add new features 2.fix some bugs

+++ Service/shop-hapi/src/main/java/com/genyuanlian/consumer/processor/BuyerConfirmOrderProcessor.java
@@ -55,7 +55,7 @@ public class BuyerConfirmOrderProcessor extends BaseApiProcessor {
 		}
 
 		OrderNoParamsVo params = new OrderNoParamsVo();
-		params.setMemberId(Integer.valueOf(memberId));
+		params.setMemberId(Long.valueOf(memberId));
 		params.setOrderNo(orderNo);
 		ShopMessageVo<String> messageVo = cardOrderApi.buyerConfirmOrder(params);
 		if (messageVo.isResult()) {
